requirements.txt
python-jose[cryptography]
passlib[bcrypt]
pydantic
fastapi
uvicorn[standard]
python-multipart
python-dotenv
cloudinary
langchain
langchain_community
langchain-groq
langchain_google_genai
pymongo
pymilvus
langgraph
duckduckgo-search


